+++ .local/bin/tag-release.sh
@@ -11,7 +11,7 @@ fi
 local_bin="`dirname "$(realpath "$0")"`"
 root="$local_bin/../.."
 
-tag="`$local_bin/get-release-tag.sh "$path"`"
+tag="v`echo -n "$(cat "$path/version")"`"
 
 description="`$local_bin/get-release-description.sh "$path"`"