+++ bin/build.sh
@@ -4,11 +4,12 @@ set -e
 
 bin="`dirname "$0"`"
 root="$bin/.."
-local_bin="$root/.local/bin"
 
 name="`basename "$(realpath "$root")"`"
+host="${FLUX_PUBLISH_DOCKER_HOST:=}"
+host_with_slash="${host}${host:+/}"
 user="${FLUX_PUBLISH_DOCKER_USER:=fluxfw}"
-image="$user/$name"
-tag="`$local_bin/get-release-tag.sh "$root"`"
+image="$host_with_slash$user/$name"
+tag="v`echo -n "$(cat "$root/version")"`"
 
 docker build "$root" --pull -t "$image:$tag" -t "$image:latest"