+++ bin/publish.sh
@@ -7,8 +7,10 @@ root="$bin/.."
 local_bin="$root/.local/bin"
 
 name="`basename "$(realpath "$root")"`"
-image="$FLUX_PUBLISH_DOCKER_USER/$name"
-tag="`$local_bin/get-release-tag.sh "$root"`"
+host="${FLUX_PUBLISH_DOCKER_HOST:=}"
+host_with_slash="${host}${host:+/}"
+image="$host_with_slash$FLUX_PUBLISH_DOCKER_USER/$name"
+tag="v`echo -n "$(cat "$root/version")"`"
 
 "$bin/build.sh"
 
@@ -19,5 +21,3 @@ export DOCKER_CONFIG="$FLUX_PUBLISH_DOCKER_CONFIG_FOLDER"
 docker push "$image:$tag"
 docker push "$image:latest"
 unset DOCKER_CONFIG
-
-"$local_bin/update-github-metadata.sh" "$root"