+++ src/Metadata/Metadata.mjs
@@ -1,3 +1,0 @@
-/**
- * @typedef {{version: string, description: string, topics: string[], homepage: string}} Metadata
- */